requirements.txt
requests==2.25.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6